# Add in-memory payload index on mmap storage (geo)