--- a/lib/segment/src/index/field_index/geo_index/immutable_geo_index.rs
+++ b/lib/segment/src/index/field_index/geo_index/immutable_geo_index.rs
@@ -212,6 +212,11 @@ impl ImmutableGeoMapIndex {
         self.points_values_count = index.points_values_count();
         self.max_values_per_point = index.max_values_per_point();
 
+        // Index is now loaded into memory, clear cache of backing mmap storage
+        if let Err(err) = index.clear_cache() {
+            log::warn!("Failed to clear mmap cache of ram mmap geo index: {err}");
+        }
+
         true
     }