Depends on <https://github.com/qdrant/qdrant/pull/6495>.

Add in-memory payload index for geo points using mmap as storage.

Loading the in-memory full text payload index from mmap is **24x faster** than from RocksDB, as shown in this simple benchmark:

- `bfb --geo-payloads -n 2000000 -d1 --indexing-threshold 100`:
    - Old RocksDB based index: 11.78s, 11.86s, 11.57s, 11.66s = avg 11.72s
    - New mmap based index: 0.489s, 0.476s, 0.476s, 0.477s = avg 0.480ms (24x faster)

### Tasks

- [x] Add test
- [x] Do load benchmark
- [x] Drop mmap caches after load
- [x] Merge <https://github.com/qdrant/qdrant/pull/6495>
- [x] Rebase on `dev`
- [x] Undraft

### All Submissions:

* [x] Contributions should target the `dev` branch. Did you create your branch from `dev`?
* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../../pulls) for the same update/change?

<!-- You can erase any parts of this template not applicable to your Pull Request. -->

### New Feature Submissions:

1. [x] Does your submission pass tests?
2. [x] Have you formatted your code locally using `cargo +nightly fmt --all` command prior to submission?
3. [x] Have you checked your code using `cargo clippy --all --all-features` command?

### Changes to Core Features:

* [x] Have you added an explanation of what your changes do and why you'd like us to include them?
* [x] Have you written new tests for your core changes, as applicable?
* [x] Have you successfully ran tests with your changes locally?